// src/phaser/types/CallsignTypes.ts
export const CarrierSpoken: { [key in CarrierCode]: string } = {
  AAL: 'American',
  ACA: 'Air Canada',
  AFR: 'Air France',
  BAW: 'Speed bird',
  ASP: 'Airsprint',
  BWA: 'Carribean',
  CMS: 'Access',
  CNK: 'Chinook',
  DAL: 'Delta',
  EGF: 'Eagleflight',
  ETH: 'Ethiopian',
  ETD: 'Etihad',
  FAB: 'First Air',
  GGN: 'Great North',
  GLR: 'Glacier',
  JAL: 'Japan Air',
  JZA: 'Jazz',
  KBA: 'Borekair',
  KEE: 'Key stone',
  LXJ: 'Flex jet',
  MAL: 'Morning star',
  MPE: 'Empress',
  NCB: 'North Cariboo',
  PAL: 'Phillipines',
  POE: 'Porter',
  PUL: 'Pulse',
  ROU: 'Rouge',
  TGO: 'Transport',
  UAL: 'Unitped',
  UPS: 'U P S',
  WEW: 'Westwynd',
  WJA: 'West jet',
};

export enum CarrierCode {
  AAL = 'AAL',
  ACA = 'ACA',
  AFR = 'AFR',
  BAW = 'BAW',
  ASP = 'ASP',
  BWA = 'BWA',
  CMS = 'CMS',
  CNK = 'CNK',
  DAL = 'DAL',
  EGF = 'EGF',
  ETH = 'ETH',
  ETD = 'ETD',
  FAB = 'FAB',
  GGN = 'GGN',
  GLR = 'GLR',
  JAL = 'JAL',
  JZA = 'JZA',
  KBA = 'KBA',
  KEE = 'KEE',
  LXJ = 'LXJ',
  MAL = 'MAL',
  MPE = 'MPE',
  NCB = 'NCB',
  PAL = 'PAL',
  POE = 'POE',
  PUL = 'PUL',
  ROU = 'ROU',
  TGO = 'TGO',
  UAL = 'UAL',
  UPS = 'UPS',
  WEW = 'WEW',
  WJA = 'WJA',
}